Added possibility of local installations in user home directory.

+++ manager/remove/remove_manager.cpp
@@ -10,9 +10,39 @@ bool RemoveManager::unlinkProducts(const std::string &prefix, const Package &pac
 {
     if (package.getLinksTo().size() > 0)
         std::cout << prefix << "removing links..." << std::endl;
+    else
+        return true;
+
+    if (!Runtime::local_folder && Runtime::try_local)
+    {
+        std::cout << "error: couldn't find the directory for local packages." << std::endl;
+
+        if (!tstd::yn_question("do you want to continue and remove it globally?"))
+        {
+            std::cout << "aborted package installation." << std::endl;
+            return false;
+        }
+    }
 
     for (int i = 0; i < package.getLinksTo().size(); i++)
     {
+        std::string to = package.getLinksTo()[i];
+
+        if (Runtime::try_local && Runtime::local_folder)
+        {
+            to = tstd::replace(to, "$usr", "~/.local");
+            to = tstd::replace(to, "$share", "~/.local/share");
+            to = tstd::replace(to, "$bin", "~/.local/bin");
+            to = tstd::replace(to, "$lib", "~/.local/lib");
+        }
+        else
+        {
+            to = tstd::replace(to, "$usr", "/usr");
+            to = tstd::replace(to, "$share", "/usr/share");
+            to = tstd::replace(to, "$bin", "/usr/bin");
+            to = tstd::replace(to, "$lib", "/usr/lib");
+        }
+
         std::cout << prefix << "unlinking " << package.getLinksTo()[i] << std::endl;
         system(std::string("if [ -f "+package.getLinksTo()[i]+" ]; then sudo unlink "+package.getLinksTo()[i]+"; fi").c_str());
     }
@@ -24,11 +54,41 @@ bool RemoveManager::removeProducts(const std::string &prefix, const Package &pac
 {
     if (package.getProductsTo().size() > 0)
         std::cout << prefix << "removing products..." << std::endl;
+    else
+        return true;
+
+    if (!Runtime::local_folder && Runtime::try_local)
+    {
+        std::cout << "error: couldn't find the directory for local packages." << std::endl;
+
+        if (!tstd::yn_question("do you want to continue and remove it globally?"))
+        {
+            std::cout << "aborted of removing the package." << std::endl;
+            return false;
+        }
+    }
 
     for (int i = 0; i < package.getProductsTo().size(); i++)
     {
-        std::cout << prefix << "removing " << package.getProductsTo()[i] << std::endl;
-        system(std::string("sudo rm -rf "+package.getProductsTo()[i]).c_str());
+        std::string to = package.getProductsTo()[i];
+
+        if (Runtime::try_local && Runtime::local_folder)
+        {
+            to = tstd::replace(to, "$usr", "~/.local");
+            to = tstd::replace(to, "$share", "~/.local/share");
+            to = tstd::replace(to, "$bin", "~/.local/bin");
+            to = tstd::replace(to, "$lib", "~/.local/lib");
+        }
+        else
+        {
+            to = tstd::replace(to, "$usr", "/usr");
+            to = tstd::replace(to, "$share", "/usr/share");
+            to = tstd::replace(to, "$bin", "/usr/bin");
+            to = tstd::replace(to, "$lib", "/usr/lib");
+        }
+
+        std::cout << prefix << "removing " << to << std::endl;
+        system(std::string("sudo rm -rf "+to).c_str());
     }
 
     return true;
@@ -39,21 +99,43 @@ void RemoveManager::uninstallPackage(const Package &p)
     std::string file = Runtime::tmp_dir+"/tmp.yaml";
     std::string url = tstd::create_url(p, "raw/master/pkg/package.yaml");
 
-    if (!tstd::download_file(url, file))
-    {
-        std::cout << "error: package " << tstd::package_to_argument(p) << " not found!" << std::endl;
-        Runtime::exit(1);
-    }
-
-    std::ofstream _of(file, std::ios::app);
-    _of << "gituser: " << p.getGitUser() << std::endl;
-    _of << "reponame: " << p.getRepoName() << std::endl;
-    _of << "server: " << p.getServer() << std::endl;
-    _of.close();
-
     std::cout << "[ remove ] now uninstalling " << tstd::package_to_argument(p) << std::endl;
 
-    Package package(YAML::LoadFile(file));
+    Package package = p;
+    YAML::Node n;
+
+    if (IPackagesManager::isPackageInstalled(p))
+        n = YAML::LoadFile(IPackagesManager::getPackage(package).getYamlPath());
+    else
+    {
+        if (!tstd::download_file(url, file))
+        {
+            std::cout << "error: package " << tstd::package_to_argument(p) << " not found!" << std::endl;
+            Runtime::exit(1);
+        }
+
+        std::ofstream _of(file, std::ios::app);
+        _of << "gituser: " << p.getGitUser() << std::endl;
+        _of << "reponame: " << p.getRepoName() << std::endl;
+        _of << "server: " << p.getServer() << std::endl;
+        _of.close();
+
+        n = YAML::LoadFile(file);
+    }
+
+    package = Package(n);
+    bool local = false;
+    bool r = false;
+
+    if (n["local"])
+        local = n["local"].as<bool>();
+
+    if (!Runtime::try_local)
+    {
+        r = true;
+        Runtime::try_local = local;
+    }
+
     if (IPackagesManager::isPackageInstalled(package))
     {
         std::string prefix = "[ " + package.getRepoName() + " ] ";
@@ -69,4 +151,7 @@ void RemoveManager::uninstallPackage(const Package &p)
     {
         std::cout << "warning: removing: package " << tstd::package_to_argument(package) << " not found. skipping." << std::endl;
     }
+
+    if (r)
+        Runtime::try_local = false;
 }
